Integrate the tuning branch in master
Tuning parameters is commonly performed in fishtest, based on an infrastructure present in a separate 'tune' branch.

However, the current situation has the following disadvantages:
- 'tune' is regularly out-of-sync with master, leading to forum posts. Keeping it up-to-date requires manual maintainer effort.
- The tuning workflow is unclear to new devs, requires more merges/commits than necessary.
- Diffs of tuning tests in fishtest are unreadable, leading to tests being approved without checking. This is a security risk, and leads to silly mistakes (e.g. approving a tuning test, which doesn't contain a TUNE()).
- The tuning branch contains unrelated changes, and is not really maintained.

Merging the branch in master fixes these issues.

There is no performance overhead of including the tuning branch:

Passed STC:
LLR: 2.96 (-2.94,2.94) [-3.00,1.00]
Total: 109710 W: 21990 L: 22032 D: 65688
http://tests.stockfishchess.org/tests/view/5add9b8c0ebc5975a378488a

Binary size is nearly unchanged (<0.5% increase).

This commit is essentially unchanged from the existing code, with small edits:
- remove unused stuff
- create tune.cpp instead of appending it somewhere
- fix license dates
- fix name (encoding) typo

No functional change.
