+++ src/ucioption.cpp
@@ -175,128 +175,3 @@ Option& Option::operator=(const string& v) {
 }
 
 } // namespace UCI
-
-
-
-/// Tuning Framework. Fully separated from SF code, appended here to avoid
-/// adding a *.cpp file and to modify Makefile.
-
-#include <iostream>
-#include <sstream>
-
-bool Tune::update_on_last;
-const UCI::Option* LastOption = nullptr;
-BoolConditions Conditions;
-static std::map<std::string, int> TuneResults;
-
-string Tune::next(string& names, bool pop) {
-
-  string name;
-
-  do {
-      string token = names.substr(0, names.find(','));
-
-      if (pop)
-          names.erase(0, token.size() + 1);
-
-      std::stringstream ws(token);
-      name += (ws >> token, token); // Remove trailing whitespace
-
-  } while (  std::count(name.begin(), name.end(), '(')
-           - std::count(name.begin(), name.end(), ')'));
-
-  return name;
-}
-
-static void on_tune(const UCI::Option& o) {
-
-  if (!Tune::update_on_last || LastOption == &o)
-      Tune::read_options();
-}
-
-static void make_option(const string& n, int v, const SetRange& r) {
-
-  // Do not generate option when there is nothing to tune (ie. min = max)
-  if (r(v).first == r(v).second)
-      return;
-
-  if (TuneResults.count(n))
-      v = TuneResults[n];
-
-  Options[n] << UCI::Option(v, r(v).first, r(v).second, on_tune);
-  LastOption = &Options[n];
-
-  // Print formatted parameters, ready to be copy-pasted in fishtest
-  std::cout << n << ","
-            << v << ","
-            << r(v).first << "," << r(v).second << ","
-            << (r(v).second - r(v).first) / 20.0 << ","
-            << "0.0020"
-            << std::endl;
-}
-
-template<> void Tune::Entry<int>::init_option() { make_option(name, value, range); }
-
-template<> void Tune::Entry<int>::read_option() {
-  if (Options.count(name))
-      value = Options[name];
-}
-
-template<> void Tune::Entry<Value>::init_option() { make_option(name, value, range); }
-
-template<> void Tune::Entry<Value>::read_option() {
-  if (Options.count(name))
-      value = Value(int(Options[name]));
-}
-
-template<> void Tune::Entry<Score>::init_option() {
-  make_option("m" + name, mg_value(value), range);
-  make_option("e" + name, eg_value(value), range);
-}
-
-template<> void Tune::Entry<Score>::read_option() {
-  if (Options.count("m" + name))
-      value = make_score(Options["m" + name], eg_value(value));
-
-  if (Options.count("e" + name))
-      value = make_score(mg_value(value), Options["e" + name]);
-}
-
-// Instead of a variable here we have a PostUpdate function: just call it
-template<> void Tune::Entry<Tune::PostUpdate>::init_option() {}
-template<> void Tune::Entry<Tune::PostUpdate>::read_option() { value(); }
-
-
-// Set binary conditions according to a probability that depends
-// on the corresponding parameter value.
-
-void BoolConditions::set() {
-
-  static PRNG rng(now());
-  static bool startup = true; // To workaround fishtest bench
-
-  for (size_t i = 0; i < binary.size(); i++)
-      binary[i] = !startup && (values[i] + int(rng.rand<unsigned>() % variance) > threshold);
-
-  startup = false;
-
-  for (size_t i = 0; i < binary.size(); i++)
-      sync_cout << binary[i] << sync_endl;
-}
-
-
-// Init options with tuning session results instead of default values. Useful to
-// get correct bench signature after a tuning session or to test tuned values.
-// Just copy fishtest tuning results in a result.txt file and extract the
-// values with:
-//
-// cat results.txt | sed 's/^param: \([^,]*\), best: \([^,]*\).*/  TuneResults["\1"] = int(round(\2));/'
-//
-// Then paste the output below, as the function body
-
-#include <cmath>
-
-void Tune::read_results() {
-
-  /* ...insert your values here... */
-}
